% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_edge}
\alias{add_edge}
\title{Add edge to graph}
\usage{
add_edge(sexp, u, v)
}
\arguments{
\item{sexp}{Graph object.}

\item{u, }{v Vertices between which an edge will be added. Must be valid vertices in the graph, and not NA.}
}
\value{
TRUE if the edge has been added, FALSE if it already existed and hasn't been added.
}
\description{
Adds an edge between vertices (u, v) if it doesn't already exist in the graph.
}
