% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{as_adj_matrix}
\alias{as_adj_matrix}
\title{Exports graph as an adjacency matrix.}
\usage{
as_adj_matrix(sexp)
}
\arguments{
\item{sexp}{Graph object.}
}
\value{
Adjacency matrix - a numeric matrix of size (V, V),
where V is the number of vertices. Element (i, j) will be set to 1
if there's an edge between vertices i, j and 0 if there isn't.
The matrix is of course square and symmetric.
}
\description{
Exports graph as an adjacency matrix.
}
