% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{from_adj_list}
\alias{from_adj_list}
\title{Creates an undirected, unweighted graph from an adjacency list.}
\usage{
from_adj_list(list)
}
\arguments{
\item{list}{Adjacency list, represented as a list of integer vectors;
k-th vector in the list represents the set of neighbors of vertex k.
Neighbors might be in any order, edges can be specified from any one direction or both,
duplicated edges will be silently ignored, leaving only a single one.
Indexes outside of range (or NA) throw an error.}
}
\value{
Graph constructed from the adjacency list.
}
\description{
Creates an undirected, unweighted graph from an adjacency list.
}
