% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{from_adj_matrix}
\alias{from_adj_matrix}
\title{Creates an undirected, unweighted graph from an adjacency matrix.}
\usage{
from_adj_matrix(matrix)
}
\arguments{
\item{matrix}{Adjacency matrix. A nonzero element at position (i, j) represent an edge between vertices (i, j).
Must be square obviously, and symmetric. NA elements will throw an error.}
}
\value{
Graph constructed from the adjacency matrix.
}
\description{
Creates an undirected, unweighted graph from an adjacency matrix.
}
