% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{from_edge_list}
\alias{from_edge_list}
\title{Creates an undirected, unweighted graph from an edgelist.}
\usage{
from_edge_list(df, V)
}
\arguments{
\item{df}{Edgelist dataframe, must have two columns: "from" and "to", representing the two vertices of each edge.
Duplicated edges are silently ignored. Indexes outside of range (or NA) throw an error.}

\item{V}{Number of vertices, must be an integer vector of length 1.}
}
\value{
Graph constructed from the edgelist.
}
\description{
Creates an undirected, unweighted graph from an edgelist.
}
