% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_igraph.R
\name{from_igraph}
\alias{from_igraph}
\title{Creates graph from igraph object.}
\usage{
from_igraph(ig)
}
\arguments{
\item{ig}{igraph object}
}
\value{
Graph object created from the igraph.
}
\description{
Creates a graph object from an igraph object.

Currently, internally uses igraph::as_edgelist and graphs::from_edge_list,
but that might change and shouldn't be relied upon.
}
