% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_graph}
\alias{new_graph}
\title{Creates an undirected, unweighted graph with the provided number of vertices.}
\usage{
new_graph(V)
}
\arguments{
\item{V}{Number of vertices, must be an integer vector of length 1.}
}
\value{
Graph object with the provided number of vertices and no edges.
}
\description{
Creates an undirected, unweighted graph with the provided number of vertices.
}
