% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remove_edge}
\alias{remove_edge}
\title{Remove edge from graph}
\usage{
remove_edge(sexp, u, v)
}
\arguments{
\item{sexp}{Graph object.}

\item{u, }{v Vertices connected by the edge which will be removed. Must be valid vertices in the graph, and not NA.}
}
\value{
TRUE if the edge has been removed, FALSE if it didn't exist and so couldn't be removed.
}
\description{
Removes an edge between vertices (u, v) if it exists in the graph.
}
