% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{triangle_count}
\alias{triangle_count}
\title{Counts triangles (cycles of length 3) in an undirected, unweighted graph}
\usage{
triangle_count(sexp, method)
}
\arguments{
\item{sexp}{The graph object, created using `new_graph` or any of the `from_*` functions}

\item{method}{The algorithm used to count triangles \itemize{
\item brute - for each edge (u, v) count how many neighbors of u are also neighbors of v, O(E * V * log(V))
\item intersect - for each edge (u, v) count the size of the intersection of their sets of neighbors, O(E * V)
}}
}
\value{
Number of unique triangles (cycles of length 3) in the graph
}
\description{
Counts triangles (cycles of length 3) in an undirected, unweighted graph
}
